refactor: readpost 패키지 분리
