+++ src/main/java/com/techfork/domain/activity/readpost/service/ReadPostQueryService.java
@@ -1,10 +1,10 @@
-package com.techfork.domain.activity.service;
+package com.techfork.domain.activity.readpost.service;
 
-import com.techfork.domain.activity.converter.ActivityConverter;
-import com.techfork.domain.activity.dto.ReadPostDto;
-import com.techfork.domain.activity.dto.ReadPostListResponse;
 import com.techfork.domain.activity.bookmark.repository.BookmarkRepository;
-import com.techfork.domain.activity.repository.ReadPostRepository;
+import com.techfork.domain.activity.readpost.converter.ReadPostConverter;
+import com.techfork.domain.activity.readpost.dto.ReadPostDto;
+import com.techfork.domain.activity.readpost.dto.ReadPostListResponse;
+import com.techfork.domain.activity.readpost.repository.ReadPostRepository;
 import com.techfork.domain.post.entity.PostKeyword;
 import com.techfork.domain.post.repository.PostKeywordRepository;
 import com.techfork.global.util.CloudflareThirdPartyThumbnailOptimizer;
@@ -22,12 +22,12 @@ import java.util.stream.Collectors;
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class ActivityQueryService {
+public class ReadPostQueryService {
 
     private final BookmarkRepository bookmarkRepository;
     private final PostKeywordRepository postKeywordRepository;
     private final ReadPostRepository readPostRepository;
-    private final ActivityConverter activityConverter;
+    private final ReadPostConverter readPostConverter;
     private final CloudflareThirdPartyThumbnailOptimizer thumbnailOptimizer;
 
     public ReadPostListResponse getReadPosts(Long userId, Long lastReadPostId, int size) {
@@ -36,7 +36,7 @@ public class ActivityQueryService {
         List<ReadPostDto> readPostsWithKeywords = attachKeywordsToReadPosts(readPosts);
         List<ReadPostDto> readPostsWithBookmarks = attachBookmarksToReadPosts(readPostsWithKeywords, userId);
 
-        return activityConverter.toReadPostListResponse(readPostsWithBookmarks, size);
+        return readPostConverter.toReadPostListResponse(readPostsWithBookmarks, size);
     }
 
     private List<ReadPostDto> attachKeywordsToReadPosts(List<ReadPostDto> readPosts) {
