+++ src/test/java/com/techfork/domain/activity/service/ActivityCommandServiceTest.java
@@ -1,15 +1,8 @@
 package com.techfork.domain.activity.service;
 
-import com.techfork.domain.activity.dto.ReadPostRequest;
 import com.techfork.domain.activity.dto.SearchHistoryRequest;
-import com.techfork.domain.activity.entity.ReadPost;
 import com.techfork.domain.activity.entity.SearchHistory;
-import com.techfork.domain.activity.repository.ReadPostRepository;
 import com.techfork.domain.activity.repository.SearchHistoryRepository;
-import com.techfork.domain.post.entity.Post;
-import com.techfork.domain.post.exception.PostErrorCode;
-import com.techfork.domain.post.repository.PostRepository;
-import com.techfork.domain.source.entity.TechBlog;
 import com.techfork.domain.useraccount.entity.User;
 import com.techfork.domain.useraccount.exception.UserErrorCode;
 import com.techfork.domain.useraccount.repository.UserRepository;
@@ -29,16 +22,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
 class ActivityCommandServiceTest {
-
-    @Mock
-    private ReadPostRepository readPostRepository;
-
-    @Mock
-    private PostRepository postRepository;
 
     @Mock
     private UserRepository userRepository;
@@ -50,149 +40,8 @@ class ActivityCommandServiceTest {
     private ActivityCommandService activityCommandService;
 
     @Test
-    @DisplayName("처음 읽는 게시글이면 조회수가 증가한다")
-    void saveReadPost_FirstRead_IncrementViewCount() {
-        // Given: 테스트 데이터 준비
-        Long userId = 1L;
-        Long postId = 100L;
-
-        User mockUser = mock(User.class);
-        TechBlog mockTechBlog = TechBlog.builder()
-                .companyName("테스트회사")
-                .blogUrl("https://test.com")
-                .rssUrl("https://test.com/rss")
-                .build();
-
-        Post mockPost = Post.builder()
-                .title("테스트 제목")
-                .fullContent("내용")
-                .plainContent("내용")
-                .company("테스트회사")
-                .url("https://test.com/post/1")
-                .publishedAt(LocalDateTime.now())
-                .crawledAt(LocalDateTime.now())
-                .techBlog(mockTechBlog)
-                .build();
-
-        ReadPostRequest request = new ReadPostRequest(
-                postId,
-                LocalDateTime.now(),
-                300
-        );
-
-        // Mock 동작 정의
-        given(userRepository.findById(userId)).willReturn(Optional.of(mockUser));
-        given(postRepository.findById(postId)).willReturn(Optional.of(mockPost));
-        given(readPostRepository.existsByUserAndPost(mockUser, mockPost)).willReturn(false); // 처음 읽음
-        given(readPostRepository.save(any(ReadPost.class))).willReturn(mock(ReadPost.class));
-
-        Long beforeViewCount = mockPost.getViewCount();
-
-        // When: saveReadPost 호출
-        activityCommandService.saveReadPost(userId, request);
-
-        // Then: 조회수가 1 증가했는지 검증
-        assertThat(mockPost.getViewCount()).isEqualTo(beforeViewCount + 1);
-
-        // 그리고 ReadPost가 저장되었는지 검증
-        verify(readPostRepository, times(1)).save(any(ReadPost.class));
-    }
-
-    @Test
-    @DisplayName("이미 읽은 게시글이면 조회수가 증가하지 않는다")
-    void saveReadPost_AlreadyRead_NoIncrementViewCount() {
-        // Given
-        Long userId = 1L;
-        Long postId = 100L;
-
-        User mockUser = mock(User.class);
-        TechBlog mockTechBlog = TechBlog.builder()
-                .companyName("테스트회사")
-                .blogUrl("https://test.com")
-                .rssUrl("https://test.com/rss")
-                .build();
-
-        Post mockPost = Post.builder()
-                .title("테스트 제목")
-                .fullContent("내용")
-                .plainContent("내용")
-                .company("테스트회사")
-                .url("https://test.com/post/1")
-                .publishedAt(LocalDateTime.now())
-                .crawledAt(LocalDateTime.now())
-                .techBlog(mockTechBlog)
-                .build();
-
-        ReadPostRequest request = new ReadPostRequest(
-                postId,
-                LocalDateTime.now(),
-                300
-        );
-
-        given(userRepository.findById(userId)).willReturn(Optional.of(mockUser));
-        given(postRepository.findById(postId)).willReturn(Optional.of(mockPost));
-        given(readPostRepository.existsByUserAndPost(mockUser, mockPost)).willReturn(true); // 이미 읽음
-        given(readPostRepository.save(any(ReadPost.class))).willReturn(mock(ReadPost.class));
-
-        Long beforeViewCount = mockPost.getViewCount();
-
-        // When
-        activityCommandService.saveReadPost(userId, request);
-
-        // Then: 조회수가 증가하지 않음
-        assertThat(mockPost.getViewCount()).isEqualTo(beforeViewCount);
-
-        // 하지만 ReadPost는 저장됨 (읽은 기록은 매번 저장)
-        verify(readPostRepository, times(1)).save(any(ReadPost.class));
-    }
-
-    @Test
-    @DisplayName("읽은 게시글 저장 실패 - 존재하지 않는 사용자")
-    void saveReadPost_Fail_UserNotFound() {
-        // Given
-        Long userId = 999L;
-        Long postId = 100L;
-        ReadPostRequest request = new ReadPostRequest(postId, LocalDateTime.now(), 300);
-
-        given(userRepository.findById(userId)).willReturn(Optional.empty());
-
-        // When & Then
-        assertThatThrownBy(() -> activityCommandService.saveReadPost(userId, request))
-                .isInstanceOf(GeneralException.class)
-                .hasFieldOrPropertyWithValue("code", UserErrorCode.USER_NOT_FOUND);
-
-        verify(userRepository, times(1)).findById(userId);
-        verify(postRepository, never()).findById(any());
-        verify(readPostRepository, never()).save(any());
-    }
-
-    @Test
-    @DisplayName("읽은 게시글 저장 실패 - 존재하지 않는 게시글")
-    void saveReadPost_Fail_PostNotFound() {
-        // Given
-        Long userId = 1L;
-        Long postId = 999L;
-        User mockUser = mock(User.class);
-        ReadPostRequest request = new ReadPostRequest(postId, LocalDateTime.now(), 300);
-
-        given(userRepository.findById(userId)).willReturn(Optional.of(mockUser));
-        given(postRepository.findById(postId)).willReturn(Optional.empty());
-
-        // When & Then
-        assertThatThrownBy(() -> activityCommandService.saveReadPost(userId, request))
-                .isInstanceOf(GeneralException.class)
-                .hasFieldOrPropertyWithValue("code", PostErrorCode.POST_NOT_FOUND);
-
-        verify(postRepository, times(1)).findById(postId);
-        verify(readPostRepository, never()).save(any());
-    }
-
-    // ===== 검색 히스토리 테스트 =====
-
-    @Test
     @DisplayName("검색 히스토리 저장 성공")
     void saveSearchHistory_Success() {
-        // Given
         Long userId = 1L;
         String query = "Spring Boot";
         LocalDateTime searchedAt = LocalDateTime.now();
@@ -202,10 +51,8 @@ class ActivityCommandServiceTest {
         given(userRepository.findById(userId)).willReturn(Optional.of(mockUser));
         given(searchHistoryRepository.save(any(SearchHistory.class))).willReturn(mock(SearchHistory.class));
 
-        // When
         activityCommandService.saveSearchHistory(userId, request);
 
-        // Then
         verify(userRepository, times(1)).findById(userId);
         ArgumentCaptor<SearchHistory> searchHistoryCaptor = ArgumentCaptor.forClass(SearchHistory.class);
         verify(searchHistoryRepository, times(1)).save(searchHistoryCaptor.capture());
@@ -215,18 +62,15 @@ class ActivityCommandServiceTest {
     @Test
     @DisplayName("검색 히스토리 저장 실패 - 존재하지 않는 사용자")
     void saveSearchHistory_Fail_UserNotFound() {
-        // Given
         Long userId = 999L;
         SearchHistoryRequest request = new SearchHistoryRequest("Spring Boot", LocalDateTime.now());
 
         given(userRepository.findById(userId)).willReturn(Optional.empty());
 
-        // When & Then
         assertThatThrownBy(() -> activityCommandService.saveSearchHistory(userId, request))
                 .isInstanceOf(GeneralException.class)
                 .hasFieldOrPropertyWithValue("code", UserErrorCode.USER_NOT_FOUND);
 
         verify(searchHistoryRepository, never()).save(any());
     }
-
 }
