+++ src/test/java/com/techfork/evaluation/recommendation/util/EvaluationFixtureLoader.java
@@ -3,8 +3,8 @@ package com.techfork.evaluation.recommendation.util;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.techfork.domain.activity.entity.ReadPost;
-import com.techfork.domain.activity.repository.ReadPostRepository;
+import com.techfork.domain.activity.readpost.entity.ReadPost;
+import com.techfork.domain.activity.readpost.repository.ReadPostRepository;
 import com.techfork.domain.post.document.ContentChunk;
 import com.techfork.domain.post.document.PostDocument;
 import com.techfork.domain.post.entity.Post;
